Add support for Spring Boot 3
Three new classes are introduced to support Spring Boot 3: StdSpringDbStateRepository, StdSpringJob and StdSpringRunner. Some new entities, repositories, and converter classes were created as part of the implementation, extending job processing functionality to work with Spring Boot 3. A module is added in the parent pom.xml file for powerimo-jobs-boot3 and the starter for Spring Boot 3 now includes this new module as a dependency.