Update pom.xml files for Spring Boot 3 Version
The commit updates the `pom.xml` files for both `powerimo-jobs-boot3` and the parent module. Spring Boot 3 version has been upgraded to 3.2.0. Also, the 'version' tag in the `powerimo-jobs-boot3` module is now set dynamically to `${revision}`. In addition, properties for `springboot3.version` and `slf4j.boot3.version` are added to the respective pom.xml.